backend/requirements.txt
absl-py==2.1.0
asgiref==3.10.0
asttokens==2.4.1
attrs==24.2.0
blinker==1.9.0
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
comtypes==1.4.8
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
Django==5.2.7
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
executing==2.0.1
Flask==3.1.0
flatbuffers==24.3.25
fonttools==4.55.0
gTTS==2.5.1
idna==3.6
ipykernel==6.29.4
ipython==8.24.0
itsdangerous==2.2.0
jax==0.4.35
jaxlib==0.4.35
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mediapipe==0.10.18
ml_dtypes==0.5.0
nest-asyncio==1.6.0
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt_einsum==3.4.0
packaging==24.0
pandas==2.2.2
parso==0.8.4
pillow==11.0.0
platformdirs==4.2.1
playsound==1.2.2
prompt-toolkit==3.0.43
protobuf==4.25.5
psutil==5.9.8
pure-eval==0.2.2
pycparser==2.22
Pygments==2.18.0
pyparsing==3.2.0
pypiwin32==223
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pyttsx3==2.98
pytz==2024.1
pywin32==306
pyzmq==26.0.3
requests==2.31.0
scikit-learn==1.5.2
scipy==1.14.1
sentencepiece==0.2.0
six==1.16.0
sounddevice==0.5.1
sqlparse==0.5.3
stack-data==0.6.3
threadpoolctl==3.5.0
tornado==6.4
traitlets==5.14.3
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.1.3
